refactor(auction-filters): 필터 UI 컴포넌트 네이밍 및 구조 정리
- FilterRadioList를 FilterRadioGroup으로 이름 변경
- FilterPriceSection을 PriceSection으로 이름 변경
- auction filter sheet 및 sidebar에서 관련 import/사용처 일괄 수정
- PriceSection의 Input 필드에 type과 aria-label 속성 추가
